todo_backend/requirements.txt
djangorestframework~=3.12
djoser~=2.1
pillow~=9.4
django~=4.1
django-cors-headers~=3.14
djangorestframework-simplejwt~=4.8
python-dotenv~=0.21
gunicorn==20.0.4
psycopg2-binary==2.8.6